fix: field price display on edit screen

+++ frontend/lib/features/admin/screens/admin_add_edit_field_screen.dart
@@ -53,8 +53,8 @@ class _AdminAddEditFieldScreenState extends State<AdminAddEditFieldScreen> {
         addressController.text = field!.address;
         typeController.text = field!.type ?? "";
         facilitiesController.text = field!.facilities ?? "";
-        priceController.text = field!.pricePerHour.toString();
-        depositController.text = field!.depositAmount.toString();
+        priceController.text = field!.pricePerHour.toInt().toString();
+        depositController.text = field!.depositAmount.toInt().toString();
         lengthController.text = field!.length?.toString() ?? "70";
         widthController.text = field!.width?.toString() ?? "50";
         grassTypeController.text = field!.grassType ?? "";
